temporal domain schema updates and controller route refinements

+++ ichaa/app/Http/Controllers/Identity/EntityAliasController.php
@@ -9,6 +9,26 @@ use App\Domain\Identity\Models\EntityAlias;
 
 class EntityAliasController extends Controller
 {
+    public function create(Entity $entity): \Illuminate\Http\RedirectResponse
+    {
+        return $this->to('entities.show', [
+            'entity' => $entity,
+            'tab' => 'aliases',
+            'compose' => 1,
+        ]);
+    }
+
+    public function edit(Entity $entity, EntityAlias $alias): \Illuminate\Http\RedirectResponse
+    {
+        abort_unless((int) $alias->entity_id === (int) $entity->id, 404);
+
+        return $this->to('entities.show', [
+            'entity' => $entity,
+            'tab' => 'aliases',
+            'edit_alias' => $alias->id,
+        ]);
+    }
+
     // POST /entities/{entity}/aliases
     public function store(Request $request, Entity $entity): \Illuminate\Http\RedirectResponse
     {
@@ -32,6 +52,8 @@ class EntityAliasController extends Controller
     // PUT /entities/{entity}/aliases/{alias}
     public function update(Request $request, Entity $entity, EntityAlias $alias): \Illuminate\Http\RedirectResponse
     {
+        abort_unless((int) $alias->entity_id === (int) $entity->id, 404);
+
         $validated = $request->validate([
             'alias'               => ['sometimes', 'string', 'max:255'],
             'alias_type'          => ['sometimes', 'string'],
@@ -50,6 +72,8 @@ class EntityAliasController extends Controller
     // DELETE /entities/{entity}/aliases/{alias}
     public function destroy(Entity $entity, EntityAlias $alias): \Illuminate\Http\RedirectResponse
     {
+        abort_unless((int) $alias->entity_id === (int) $entity->id, 404);
+
         $alias->delete();
 
         return $this->back('Alias removed.');
